[Add] Bookmark非同期化　一覧未完

+++ app/controllers/public/bookmarks_controller.rb
@@ -4,15 +4,25 @@ class Public::BookmarksController < ApplicationController
 
   def create
     @post = Post.find(params[:post_id])
-  	bookmark = current_user.bookmarks.new(post_id: @post.id)
-  	bookmark.save
-  	redirect_back(fallback_location: root_path)
+    bookmark = current_user.bookmarks.new(post_id: @post.id)
+    if Bookmark.exists?(post_id: @post.id, user_id: current_user.id)
+      render :create
+    else
+      bookmark.save
+      render :create
+    end
+
   end
 
   def destroy
     @post = Post.find(params[:post_id])
-  	bookmark = current_user.bookmarks.find_by(post_id: @post.id)
-  	bookmark.destroy
-  	redirect_back(fallback_location: root_path)
+    bookmark = current_user.bookmarks.find_by(post_id: @post.id)
+    unless bookmark == nil
+      bookmark.destroy
+      render :destroy
+    else
+      render :destroy
+    end
+
   end
 end
